Added Some Javadoc
Added back in the missing tooltips.
