+++ src/main/java/com/InfinityRaider/AgriCraft/items/blocks/ItemBlockGrate.java
@@ -1,0 +1,61 @@
+package com.InfinityRaider.AgriCraft.items.blocks;
+
+import java.util.List;
+
+import net.minecraft.block.Block;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.StatCollector;
+import net.minecraft.world.World;
+import net.minecraftforge.common.util.ForgeDirection;
+
+import com.InfinityRaider.AgriCraft.tileentity.decoration.TileEntityGrate;
+import com.InfinityRaider.AgriCraft.utility.PlayerHelper;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
+
+public class ItemBlockGrate extends ItemBlockCustomWood {
+    public ItemBlockGrate(Block block) {
+        super(block);
+    }
+    
+    @Override
+    @SideOnly(Side.CLIENT)
+    public void addMoreInformation(ItemStack stack, List list) {
+    	//All this just to add a single line...
+        list.add(StatCollector.translateToLocal("agricraft_tooltip.grate"));
+    }
+
+    @Override
+    public boolean placeBlockAt(ItemStack stack, EntityPlayer player, World world, int x, int y, int z, int side, float hitX, float hitY, float hitZ, int metadata) {
+        if (super.placeBlockAt(stack, player, world, x, y, z, side, hitX, hitY, hitZ, metadata)) {
+            TileEntityGrate grate = (TileEntityGrate) world.getTileEntity(x, y, z);
+            if (side == 0 || side == 1) {
+                ForgeDirection dir = PlayerHelper.getPlayerFacing(player);
+                if (dir == ForgeDirection.NORTH || dir == ForgeDirection.SOUTH) {
+                    setOffsetAndOrientation(grate, hitZ, (short) 0);
+                } else {
+                    setOffsetAndOrientation(grate, hitX, (short) 1);
+                }
+            } else {
+                setOffsetAndOrientation(grate, hitY, (short) 2);
+            }
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    private void setOffsetAndOrientation(TileEntityGrate grate, float hit, short orientation) {
+        grate.setOrientationValue(orientation);
+        if (hit <= 0.3333F) {
+            grate.setOffSet((short) 0);
+        } else if (hit <= 0.6666F) {
+            grate.setOffSet((short) 1);
+        } else {
+            grate.setOffSet((short) 2);
+        }
+        grate.calculateBounds();
+    }
+}
